Moved all appcd command files into a commands directory.
